Category

+++ database/migrations/2025_04_24_121046_create_setting_translations_table.php
@@ -14,14 +14,16 @@ class CreateSettingTranslationsTable extends Migration
     public function up()
     {
 
+     if(!Schema::hasTable('setting_translations'))   {
             Schema::create('setting_translations', function (Blueprint $table) {
-               $table->increments('id');
-               $table->integer('setting_id')->unsigned();
-               $table->string('locale');
-               $table->longText('value');
-               $table->unique(['setting_id', 'locale']);
-               $table->foreign('setting_id')->references('id')->on('settings')->onDelete('cascade');
+                $table->increments('id');
+                $table->integer('setting_id')->unsigned();
+                $table->string('locale');
+                $table->longText('value');
+                $table->unique(['setting_id', 'locale']);
+                $table->foreign('setting_id')->references('id')->on('settings')->onDelete('cascade');
             });
+        }
 
     }
